Changed int8_t to int in order to fix Xcode 9 "undefined behavior" build error.

+++ ZXingObjC/common/ZXByteArray.m
@@ -31,7 +31,7 @@
   return self;
 }
 
-- (id)initWithBytes:(int8_t)byte1, ... {
+- (id)initWithBytes:(int)byte1, ... {
   va_list args;
   va_start(args, byte1);
   unsigned int length = 0;
